+++ RELEASE-NOTES.md
@@ -45,6 +45,7 @@ Project Page: https://github.com/a2stuff/a2d
 * Add Apple+Control+W as shortcut to focus window.
 * Fix behavior after File > Duplicate failed due to space or storage type.
 * Set empty window maprect so that when restored with new file scrollbars are inactive.
+* Fix volume icon flicker after Check All Drives for removable disks.
 
 ### Selector